Enhancement One, logging and Tests added
Refactor the utility tool to not write to the log file but instead only the console. Enhance the logic of this to utilize console.log(), console.error() and console.warn() respectively. This may require refactoring of their unit tests depending on your adjustments.

Refactor the utility tool to not write to the log file but instead only the console. Enhance the logic of this to utilize console.log(), console.error() and console.warn() respectively.